% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRofexPosition.R
\name{.secuencia}
\alias{.secuencia}
\title{.secuencia}
\usage{
.secuencia(serie, cal)
}
\arguments{
\item{serie}{Serie de fechas, separadas por 1 mes para generar la secuencia. e.g.: "1970-01-25" "1970-02-25" "1970-03-25"}
}
\value{
description Un tibble con futuro y vto.
}
\description{
Función interna del paquete Rofex que genera la secuencia de posiciones a consultar a la API de Rofex.
}
