% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRofexPosition.R
\name{getRofexCurrentCurve}
\alias{getRofexCurrentCurve}
\title{getRofexCurrentCurve}
\usage{
getRofexCurrentCurve()
}
\description{
Devuelve la curva desde la posición actual. Mes actual más 11 posiciones
}
