% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRofexPosition.R
\name{getRofexPosition}
\alias{getRofexPosition}
\title{getRofexPosition}
\usage{
getRofexPosition(
  position = "curva",
  from = Sys.Date() - 1,
  to = Sys.Date(),
  page = 1,
  pageSize = 32000
)
}
\arguments{
\item{position}{La/s posiciones a consultar.}

\item{from}{Fecha desde}

\item{to}{Fecha hasta}

\item{page}{Numero de página a devolver según paginado}

\item{pageSize}{Cantidad de valores incluídos en cada páginado}
}
\value{
Devuelve una lista. El primer elemento es una tibble con los datos de las posiciones consultadas.
El segundo elemento es una tibble con las posiciones que fallaron.
}
\description{
Trae una o varias posiciones desde la API de Rofex con toda la información de Rofex
más el día que vence cada contrato según el calendario cargado en "~/data/test.sqlite3"
}
