% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRofexPosition.R
\name{posiciones}
\alias{posiciones}
\title{posiciones}
\usage{
posiciones(start_month, start_year)
}
\arguments{
\item{start_month}{Mes de inicio}

\item{start_year}{Año de inicio}
}
\value{
Un vector con las posiciones de Rofex a consultar
}
\description{
Funcion que devuelve las posiciones de rofex a partir de un mes y un año dados en formato DLRMMYYYY
}
